% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors_from_image.R
\name{get_colors_from_image}
\alias{get_colors_from_image}
\title{The get_colors_from_image function}
\usage{
get_colors_from_image(image, n = 5, order_by = "luminance")
}
\arguments{
\item{image}{The path to the image. Supports most image types.
See image_read from magick package.}

\item{n}{The number of colors in the palette. Defaults to 5.}

\item{order_by}{How to order the colors. One of 'luminance', 'chroma' or 'hue'. Defaults to 'luminance'}
}
\value{
The hex codes of the n colors in the palette
}
\description{
This function allows you to identify the dominant colors in an image.
}
\examples{
get_colors_from_image("https://raw.githubusercontent.com/HughSt/mappalettes/master/images/nathan-lindahl-1j18807_ul0-unsplash.jpg",5)
}
