% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin_points.R
\name{hexbin_points}
\alias{hexbin_points}
\title{The hexbin_points function}
\usage{
hexbin_points(points, n, fun, z, buffer = 0.1, return.na = FALSE)
}
\arguments{
\item{points}{The sfc points object to summarize}

\item{n}{The approximate number of hexbin cells}

\item{fun}{The function to apply when summarizing point values in each hexbin cell}

\item{z}{The column of the SF points object to apply the function to}

\item{buffer}{The distance by which to buffer the convex hull
of points used to create the hexbin cells as a proportion of the
maximum interpoint distance. By default set to 0.1 (10\%)}

\item{return.na}{Logical. Whether or not to return hexbin cells with NAs.}
}
\description{
This function allows you to create a hexbin plot from point data.
}
\examples{
# Load BCG vaccination data
data("BCG_vaccination_DRC")

# Get mean coverage per hexbin
hexbin_points(BCG_vaccination_DRC, n = 250, fun = mean, z = "coverage")

# Get number of points per hexbin
hexbin_points(BCG_vaccination_DRC, n = 250, fun = length, z = "coverage")
}
\keyword{hexbin}
