% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin_raster.R
\name{hexbin_raster}
\alias{hexbin_raster}
\title{The hexbin_raster function}
\usage{
hexbin_raster(r, n, fun)
}
\arguments{
\item{r}{The rasterLayer}

\item{n}{The approximate number of hexbin cells}

\item{fun}{The function you want to apply (e.g. mean, max, min) when summarizing raster values in each hexbin}
}
\description{
This function allows you to create a hexbin plot of a rasterLayer.
}
\examples{
hexbin_raster("swz_elev", 100, mean)
}
\keyword{hexbin}
