% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappalettes_function.R
\name{map_palette}
\alias{map_palette}
\title{The map palette function}
\usage{
map_palette(name, n = 5)
}
\arguments{
\item{name}{The name of the palette}

\item{n}{The number of colors in the palette}
}
\description{
This function allows you to easily access beautiful palettes designed for maps.
}
\keyword{cats}
