% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin_raster.R
\name{hexbin_raster}
\alias{hexbin_raster}
\title{The hexbin_raster function}
\usage{
hexbin_raster(r, n, fun)
}
\arguments{
\item{r}{The rasterLayer}

\item{n}{The approximate number of hexbin cells}

\item{fun}{The function you want to apply when summarizing raster values in each hexbin.
Note that hexbin_raster uses velox which currently does not support na.rm. To ignore NAs
pass function which includes na.rm. See examples.}
}
\description{
This function allows you to create a hexbin plot of a rasterLayer.
}
\examples{
# Generate hexbins and calculate raster mean in each bin
hexbin_raster("swz_elev", 100, function(x) mean(x, na.rm=TRUE))
}
\keyword{hexbin}
