% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpalettes_function.R
\name{view_palette}
\alias{view_palette}
\title{The map palette function}
\usage{
view_palette(name, n = 5)
}
\arguments{
\item{name}{The name of the palette}

\item{n}{The number of colors in the palette}
}
\description{
This function allows you to easily access beautiful palettes designed for maps.
}
\examples{
view_palette("green_machine")
}
\keyword{cats}
