% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewpalettes_function.R
\name{view_palette}
\alias{view_palette}
\title{The view palette function}
\usage{
view_palette(palette, n = 5, type = "bars")
}
\arguments{
\item{palette}{The name of the palette. One of "green_machine",
"irish_flag", "tealberry_pie", "sunset", "the_joker" or "bruiser"
or a vector of 5 hex codes.}

\item{n}{The number of colors in the palette, Defaults to 5. Ignored if hex codes
provided as `palette`}

\item{type}{Either "bars" which shows palette as bars,
"raster" which shows elevation in Swaziland or "polys" which
shows mean elevation in Swaziland by admin 2 area.}
}
\description{
This function allows you to easily view beautiful palettes designed for maps.
}
\examples{
view_palette("bruiser")

# Use custom set of hexcodes
view_palette(c("#1D3141","#096168","#209478","#75C56E","#E2EE5E"),
type="raster")
}
